reproduces the panic, although there's a lot of cruft I can probably remove.
Much of the code here I lifted from mb2, so there are additional
stores that probably aren't needed.  I have, however, taken out stuff
and then fixed the remaining codd to compile, which means that what is
here is non-sensical.

It should be easy for me to pare this down into something smaller.